requirements.txt
pandas>=2.2.0
numpy>=2.0.0
scikit-learn>=1.4.0
lightgbm>=4.1.0
PyYAML>=6.0
pytest>=7.1.2
jupyter>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.64.0
python-dotenv>=0.20.0
click>=8.1.3